Add files via upload
This project was created for a course submission at the end of the term. I made use of the potentiometer in my Arduino board (grove board) as a volume controller for my desktop. I wrote both the Arduino and java codes since we were required to showcase the communication between the board and the 2 IDEs in the project. There is also a PDF file inside the zip file that shows everything there is to learn about this project.
Here is the link to the video that showcases my journey in developing this project: https://drive.google.com/file/d/1K1KgWFERFm-qcTzyi14dNNFSHN3IJNHM/view